requirements.txt
amqp==2.3.2
aniso8601==3.0.2
autopep8==1.3.5
Babel==2.6.0
backports-abc==0.5
billiard==3.5.0.4
celery==4.2.1
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.36.0
django-appconf==1.0.2
django-autoslug==1.9.3
django-cacheops==4.0.7
django-cors-headers==2.4.0
django-dbbackup==3.2.0
django-debug-toolbar==1.9.1
django-elasticsearch-dsl==0.5.0
django-filter==1.1.0
django-letsencrypt==3.0.1
django-redis==4.9.0
django-storages==1.6.6
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
flower==0.9.2
funcy==1.10.3
gprof2dot==2016.10.13
graphene==2.1.2
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
idna==2.7
ipaddress==1.0.22
iso8601==0.1.12
Jinja2==2.10
kombu==4.2.1
litresapi==1.0.1
lxml==4.2.4
MarkupSafe==1.0
MonthDelta==0.9.1
numpy==1.15.0
oauthlib==2.1.0
olefile==0.45.1
pilkit==2.0
Pillow==5.2.0
promise==2.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.4.0
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.3
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
requests-oauthlib==1.0.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.11.0
sorl-thumbnail==12.4.1
sqlparse==0.2.4
tornado==5.1
typing==3.6.4
urllib3==1.23
vine==1.1.4
vk==2.0.2
